+++ app/(auth)/project/[link]/page.tsx
