+++ app/loading.tsx
@@ -1,0 +1,25 @@
+"use client"
+
+import React from 'react'
+import LoadingBar from 'react-top-loading-bar'
+
+
+const Loading = () => {
+  const [progress, setProgress] = React.useState(0)
+
+  React.useEffect(() => {
+    setProgress(400)
+  }, [])
+  return (
+    <div>
+      <LoadingBar
+        color='#f11946'
+        progress={progress}
+        onLoaderFinished={() => setProgress(0)}
+        waitingTime={1000}
+      />
+    </div>
+  )
+}
+
+export default Loading
