+++ components/Project/SingleProject.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function SingleProject() {
+  return (
+    <div>SingleProject</div>
+  )
+}
