+++ helper/jwtDecode.ts
@@ -2,7 +2,7 @@ import { jwtDecode } from "jwt-decode";
 
 export const jwtDecoder = (token:any) => {
   if (token) {
-    const tokenDecoded = jwtDecode(token);
+    const tokenDecoded = jwtDecode(token || null);
     return tokenDecoded;
   }
 };
